Update handling of server side includes
ensure quotes are handled correctly when server side
includes are added to attributes
