+++ Manifest.txt
@@ -235,3 +235,4 @@ lib/xml-apis.jar
 lib/xsd/xmlparser/nokogiri.rb
 patches/libxml2/0001-Revert-Do-not-URI-escape-in-server-side-includes.patch
 patches/libxml2/0002-Remove-script-macro-support.patch
+patches/libxml2/0003-Update-entities-to-remove-handling-of-ssi.patch
